# docs/api/openapi.yaml
openapi: '3.0.3'
info:
  title: Kalki-Mirror API
  version: 0.2.1
  description: >-
    REST API for the KALKI Tantrik Intelligence Platform.
    Pattern recognition, transit geometry, AI-powered Vedic content,
    and membership management.
  contact:
    name: KALKI Platform
    url: https://www.astrokalki.com
  license:
    name: Proprietary
servers:
  - url: https://www.astrokalki.com
    description: Production
...
paths:
...
  /api/admin/warroom:
    get:
      operationId: apiAdminWarroomGet
      tags: [Admin]
      summary: WAR ROOM API — campaign intelligence for the attribution layer.
      security:
        - adminAuth: []
      responses:
        '200':
          description: OK (shape documented in the route module)
        '401':
          description: Unauthorized — admin session required
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT via Bearer token (cookie-based).
        Used for standard authenticated routes.
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        NextAuth JWT with ADMIN, SUPERADMIN, EDITOR, or REVIEWER role required.

  # ── Reusable Responses ──────────────────────────────────
...